Separate parser and tokenizer tests
